+++ src/components/ActionBar/tests/ActionBar.stories.tsx
@@ -1,5 +1,10 @@
+import { StoryObj } from "@storybook/react-vite";
+import { expect } from "storybook/test";
 import { Placeholder, Example } from "utilities/storybook";
 import ActionBar from "components/ActionBar";
+import View from "components/View";
+import Button from "components/Button";
+import useToggle from "hooks/useToggle";
 
 export default {
 	title: "Components/ActionBar",
@@ -11,8 +16,24 @@ export default {
 	},
 };
 
-export const position = {
-	name: "position",
+const Fixtures = {
+	Container: (props: React.PropsWithChildren) => (
+		<View backgroundColor="neutral-faded" height="160px" overflow="hidden" borderRadius="medium">
+			{props.children}
+		</View>
+	),
+	Actions: () => (
+		<View direction="row" gap={2}>
+			<Button onClick={() => {}}>Action</Button>
+			<Button onClick={() => {}} variant="outline">
+				Action
+			</Button>
+		</View>
+	),
+};
+
+export const positionRelative = {
+	name: "position, positionType: relative",
 	render: () => (
 		<Example>
 			<Example.Item title="position: top">
@@ -22,11 +43,99 @@ export const position = {
 			</Example.Item>
 
 			<Example.Item title="position: bottom">
-				<ActionBar>
+				<ActionBar position="bottom">
 					<Placeholder />
 				</ActionBar>
 			</Example.Item>
 		</Example>
+	),
+};
+
+export const positionAbsolute = {
+	name: "position, positionType: absolute",
+	render: () => (
+		<Example>
+			<Example.Item title="position: top-start">
+				<Fixtures.Container>
+					<ActionBar padding={2} position="top-start" positionType="absolute">
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+
+			<Example.Item title="position: top">
+				<Fixtures.Container>
+					<ActionBar padding={2} position="top" positionType="absolute">
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+
+			<Example.Item title="position: top-end">
+				<Fixtures.Container>
+					<ActionBar padding={2} position="top-end" positionType="absolute">
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+
+			<Example.Item title="position: bottom-start">
+				<Fixtures.Container>
+					<ActionBar padding={2} position="bottom-start" positionType="absolute">
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+
+			<Example.Item title="position: bottom">
+				<Fixtures.Container>
+					<ActionBar padding={2} position="bottom" positionType="absolute">
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+
+			<Example.Item title="position: bottom-end">
+				<Fixtures.Container>
+					<ActionBar padding={2} position="bottom-end" positionType="absolute">
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+		</Example>
+	),
+};
+
+export const positionFixed = {
+	name: "position, positionType: fixed",
+	render: () => (
+		<>
+			<ActionBar padding={2} position="top-start" positionType="fixed">
+				<Fixtures.Actions />
+			</ActionBar>
+
+			<ActionBar padding={2} position="top" positionType="fixed">
+				<Fixtures.Actions />
+			</ActionBar>
+
+			<ActionBar padding={2} position="top-end" positionType="fixed">
+				<Fixtures.Actions />
+			</ActionBar>
+
+			<ActionBar padding={2} position="bottom-start" positionType="fixed">
+				<Fixtures.Actions />
+			</ActionBar>
+
+			<ActionBar padding={2} position="bottom" positionType="fixed">
+				<Fixtures.Actions />
+			</ActionBar>
+
+			<ActionBar padding={2} position="bottom-end" positionType="fixed">
+				<Fixtures.Actions />
+			</ActionBar>
+
+			<div style={{ height: 2000 }} />
+		</>
 	),
 };
 
@@ -44,6 +153,78 @@ export const elevated = {
 				<ActionBar elevated>
 					<Placeholder />
 				</ActionBar>
+			</Example.Item>
+
+			<Example.Item title="auto elevated, position: bottom">
+				<Fixtures.Container>
+					<ActionBar position="bottom-end">
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+		</Example>
+	),
+};
+
+export const offset = {
+	name: "offset",
+	render: () => (
+		<Example>
+			<Example.Item title="offset 2, position: top">
+				<Fixtures.Container>
+					<ActionBar position="top" positionType="absolute" offset={2}>
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+
+			<Example.Item title="offset 2, position: bottom-end">
+				<Fixtures.Container>
+					<ActionBar position="bottom-end" offset={2}>
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+
+			<Example.Item title="offset s: 2, m: 4, position: bottom-end">
+				<Fixtures.Container>
+					<ActionBar position="bottom-end" offset={{ s: 2, m: 4 }}>
+						<Fixtures.Actions />
+					</ActionBar>
+				</Fixtures.Container>
+			</Example.Item>
+		</Example>
+	),
+};
+
+export const active = {
+	name: "active",
+	render: () => {
+		const barToggle = useToggle();
+
+		return (
+			<>
+				<Button onClick={() => barToggle.toggle()}>Toggle</Button>
+				<ActionBar active={barToggle.active} positionType="fixed" position="top-end">
+					<Fixtures.Actions />
+				</ActionBar>
+			</>
+		);
+	},
+};
+
+export const blurred = {
+	name: "blurred",
+	render: () => (
+		<Example>
+			<Example.Item title="blurred">
+				<View backgroundColor="neutral-faded" height="200px" align="end" justify="end" padding={8}>
+					<Button color="primary">Action</Button>
+
+					<ActionBar position="bottom-end" blurred>
+						<View width={20} height={20} />
+					</ActionBar>
+				</View>
 			</Example.Item>
 		</Example>
 	),
@@ -72,4 +253,21 @@ export const padding = {
 			</Example.Item>
 		</Example>
 	),
+};
+
+export const className: StoryObj = {
+	name: "className, attributes",
+	render: () => (
+		<div data-testid="root">
+			<ActionBar className="test-classname" attributes={{ id: "test-id" }}>
+				<Placeholder />
+			</ActionBar>
+		</div>
+	),
+	play: async ({ canvas }) => {
+		const root = canvas.getByTestId("root").firstChild;
+
+		expect(root).toHaveClass("test-classname");
+		expect(root).toHaveAttribute("id", "test-id");
+	},
 };