+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 3.7.0
 
+- ActionBar: Added active, offset, more positions, positionType, blurred props
+
 - Added Attributes type to the package exports
 - Added `mono` font family to the themes
 - Added zIndex tokens to the themes, updated components to use new zIndex tokens